style: refactor some comments and variable names

+++ app/src/main/java/com/android/shelfLife/model/foodFacts/FoodFactsRepository.kt
@@ -2,7 +2,7 @@ package com.android.shelfLife.model.foodFacts
 
 // Sealed class to represent different types of search inputs
 sealed class FoodSearchInput {
-    data class Barcode(val barcode: Int) : FoodSearchInput()
+    data class Barcode(val barcode: Long) : FoodSearchInput()
     data class Query(val searchQuery: String) : FoodSearchInput()
 }
